Add City Access Map
CityAccessMap was developed by urban scientists, for urban planners, with the objective to drastically simplify this task. It uses open data from OpenStreetMap and the Global Human Settlement Layer (GHSL) by the European Commission, to quantify and visualize walking accessibility to a variety of services for any city in the world.